+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * [#413](https://github.com/ruby-grape/grape-swagger/pull/413): Move all model parsing logic to separate gems `grape-swagger-entity` and added representable parser `grape-swagger` - [@Bugagazavr](https://github.com/Bugagazavr).
 * Your contribution here.
 
+#### Fixes
+
+* [#416](https://github.com/ruby-grape/grape-swagger/pull/416): Support recursive models - [@lest](https://github.com/lest).
+
 ### 0.20.3 (May 9, 2016)
 
 #### Features